RAG/requirements.txt
flask
flask-cors
python-dotenv
langchain-community
langchain-core
langchain-milvus
langchain-text-splitters
pymilvus
openai
dashscope
pypdfium2
